Refactor UserJdbcRepositoryImpl to use mappers for ResultSet conversion
- Add new map methods in AddressMapper and UserMapper to convert ResultSet to entities
- Update UserJdbcRepositoryImpl to use the new mapper methods
- Simplify the findAll method by leveraging the new mapper functionality
- Improve code readability and maintainability by separating mapping logic
